feat: stream persistence
