+++ packages/server-ai/src/ai/thread.controller.ts
@@ -46,7 +46,7 @@ import { AiService } from './ai.service'
 import { RunCreateStreamCommand, ThreadCreateCommand, ThreadDeleteCommand } from './commands'
 import { FindThreadQuery, SearchThreadsQuery } from './queries'
 import type { components } from './schemas/agent-protocol-schema'
-import { RedisSseStreamService, SseLockOwnerCandidate, SseLockSnapshot } from './stream/redis-sse.service'
+import { RedisSseStreamService, SseLockOwnerCandidate, SseLockSnapshot } from '../shared/stream'
 import { CancelConversationCommand } from '../chat-conversation'
 import { GetChatConversationQuery } from '../chat-conversation'
 import { CopilotUserUsageQuery } from '../copilot-user/queries'
