+++ packages/server-ai/src/shared/index.ts
@@ -7,5 +7,6 @@ export * from './volume/index'
 export * from './file-upload-targets'
 export * from './tools/index'
 export * from './execution/index'
+export * from './stream/index'
 export * from './super-admin-organization-scope.module'
 export * from './super-admin-organization-scope.service'
