+++ packages/server-ai/src/xpert/commands/chat.command.ts
@@ -2,19 +2,26 @@ import { TChatOptions, TChatRequest } from '@xpert-ai/contracts'
 import { Command } from '@nestjs/cqrs'
 import { Observable } from 'rxjs'
 
-export class XpertChatCommand extends Command<Observable<MessageEvent>> {
-	static readonly type = '[Xpert] Chat'
+export interface XpertChatStreamPersistenceOptions {
+    transport: 'redis-stream'
+    threadId?: string | null
+    runId?: string | null
+}
 
-	constructor(
-		public readonly request: TChatRequest,
-		public readonly options?: TChatOptions & {
-			xpertId?: string
-			// Use xpert's draft
-			isDraft?: boolean
-			fromEndUserId?: string
-			execution?: {id: string}
-		}
-	) {
-		super()
-	}
+export class XpertChatCommand extends Command<Observable<MessageEvent>> {
+    static readonly type = '[Xpert] Chat'
+
+    constructor(
+        public readonly request: TChatRequest,
+        public readonly options?: TChatOptions & {
+            xpertId?: string
+            // Use xpert's draft
+            isDraft?: boolean
+            fromEndUserId?: string
+            execution?: { id: string }
+            streamPersistence?: XpertChatStreamPersistenceOptions
+        }
+    ) {
+        super()
+    }
 }
